backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
websockets==12.0
python-dotenv==1.0.1
reportlab==4.1.0
pydantic==2.7.1
